Modification js

+++ app/Http/Controllers/OutilController.php
@@ -37,7 +37,6 @@ class OutilController extends Controller
         }
         $list = $lists->paginate(10);
         return view('viewadmindste.outils.listoutils', compact('list'));
-       
     }
 
     public function getallchamp(Request $request)
@@ -129,7 +128,7 @@ class OutilController extends Controller
             ->get(['code', 'libelle']);
         return response()->json($libelles);
     }
-    
+
     public function affectUserInOutil(Request $request)
     {
         try {
@@ -223,9 +222,9 @@ class OutilController extends Controller
                 return view("vendor.error.649");
             } else {
                 return json_encode(["data" => DB::table('traces')
-                ->join('utilisateurs', "utilisateurs.idUser", "=", "traces.action")
-                ->select('nom', 'prenom', 'libelle', 'traces.created_at as created_at')
-                ->where('type', "outil")->where("idsec", $request->id)->get()]);
+                    ->join('utilisateurs', "utilisateurs.idUser", "=", "traces.action")
+                    ->select('nom', 'prenom', 'libelle', 'traces.created_at as created_at')
+                    ->where('type', "outil")->where("idsec", $request->id)->get()]);
             }
         } catch (\Exception $e) {
             return Back()->with('error', "Une erreur ses produites :" . $e->getMessage());
@@ -343,17 +342,28 @@ class OutilController extends Controller
             if (!in_array("delete_outil", session("auto_action"))) {
                 return view("vendor.error.649");
             } else {
-                $lib = Outil::where('id', request('id'))->first()->nameoutils;
-                $occurence = json_encode(Outil::where('id', request('id'))->first());
+                $outils = Outil::find(request('id'));
+                // dd($outils);
+                if ($outils) {
+                    $lib = Outil::where('id', request('id'))->first()->nameoutils;
+                    $occurence = json_encode(Outil::where('id', request('id'))->first());
 
-                TraceController::setTrace("Data delete : " . $occurence, session("utilisateur")->idUser);
+                    TraceController::setTrace("Data delete : " . $occurence, session("utilisateur")->idUser);
 
-                Outil::where('id', request('id'))->delete();
-                $info = "Vous avez supprimé " . $lib . " avec succès.";
-                return $info;
+                    Outil::where('id', request('id'))->delete();
+                    $info = "Vous avez supprimé l'outils " . $lib . " avec succès.";
+                    flash($info)->success();
+                    return $info;
+                } else {
+                    $info = "Outils introuvable.";
+                    flash($info)->error();
+                    return $info;
+                }
             }
         } catch (\Exception $e) {
-            return Back()->with('error', "Une erreur ses produites :" . $e->getMessage());
+            $errorString = "Une erreur ses produites" .  $e->getMessage();
+            flash("Erreur : " . $errorString)->error();
+            return $errorString;
         }
     }
 
@@ -470,67 +480,64 @@ class OutilController extends Controller
 
 
     //outils export
-	public function exportoutils(Request $request)
-	{
-		try {
+    public function exportoutils(Request $request)
+    {
+        try {
 
             $list = Outil::orderBy("categorie", "asc")->get();
-           
-			$format = $request->format;
-	
-			// Générer le fichier en fonction du format demandé
-			switch ($format) {
-				case 'pdf':
-					$pdfExporter = new OutilspdfExport($list);
-					$filePath = $pdfExporter->generatePdf();
-					$pdfContent = Storage::get($filePath);
-	
-					return response($pdfContent, 200)
-							->header('Content-Type', 'application/pdf')
-							->header('Content-Disposition', 'attachment; filename="OutilsExport.pdf"');
-				case 'xlsx':
-				default:
-					return Excel::download(new OutilsExport($list), 'OutilsExport.xlsx', \Maatwebsite\Excel\Excel::XLSX);
-			}
-		} catch (\Exception $e) {
-			return response()->json(["status" => 1, "message" => "Erreur lors du téléchargement : " . $e->getMessage()], 400);
-		}
-	}
 
-
-	//export outils historique
-	public function expoutilhisto(Request $request)
-	{
-		try 
-        {
-            $data = DB::table('traces')
-                    ->join('utilisateurs', 'utilisateurs.idUser', '=', 'traces.action') // Correction ici
-                    ->join('outils', 'outils.id', '=', 'traces.idsec') // Si vous voulez inclure la table 'outils'
-                    ->select('nom', 'prenom', 'libelle', 'traces.created_at as created_at', 'outils.nameoutils')
-                    ->where('traces.type', 'outil')
-                    ->where('traces.idsec', $request->idhisto)
-                    ->get();
-             
             $format = $request->format;
-	
+
+            // Générer le fichier en fonction du format demandé
+            switch ($format) {
+                case 'pdf':
+                    $pdfExporter = new OutilspdfExport($list);
+                    $filePath = $pdfExporter->generatePdf();
+                    $pdfContent = Storage::get($filePath);
+
+                    return response($pdfContent, 200)
+                        ->header('Content-Type', 'application/pdf')
+                        ->header('Content-Disposition', 'attachment; filename="OutilsExport.pdf"');
+                case 'xlsx':
+                default:
+                    return Excel::download(new OutilsExport($list), 'OutilsExport.xlsx', \Maatwebsite\Excel\Excel::XLSX);
+            }
+        } catch (\Exception $e) {
+            return response()->json(["status" => 1, "message" => "Erreur lors du téléchargement : " . $e->getMessage()], 400);
+        }
+    }
+
+
+    //export outils historique
+    public function expoutilhisto(Request $request)
+    {
+        try {
+            $data = DB::table('traces')
+                ->join('utilisateurs', 'utilisateurs.idUser', '=', 'traces.action') // Correction ici
+                ->join('outils', 'outils.id', '=', 'traces.idsec') // Si vous voulez inclure la table 'outils'
+                ->select('nom', 'prenom', 'libelle', 'traces.created_at as created_at', 'outils.nameoutils')
+                ->where('traces.type', 'outil')
+                ->where('traces.idsec', $request->idhisto)
+                ->get();
+
+            $format = $request->format;
+
             // Générer le fichier en fonction du format demandé
             switch ($format) {
                 case 'pdf':
                     $pdfExporter = new OutilhistopdfExport($data);
                     $filePath = $pdfExporter->generatePdf();
                     $pdfContent = Storage::get($filePath);
-                
+
                     return response($pdfContent, 200)
-                    ->header('Content-Type', 'application/pdf')
-                    ->header('Content-Disposition', 'attachment; filename="HistoriqueOutilsExport.pdf"');
+                        ->header('Content-Type', 'application/pdf')
+                        ->header('Content-Disposition', 'attachment; filename="HistoriqueOutilsExport.pdf"');
                 case 'xlsx':
                 default:
                     return Excel::download(new HistoExport($data), 'HistoriqueOutilsExport.xlsx', \Maatwebsite\Excel\Excel::XLSX);
             }
-
-		} catch (\Exception $e) {
-			return response()->json(["status" => 1, "message" => "Erreur lors du telechargement : " . $e->getMessage()], 400);
-		}
-	}
-
+        } catch (\Exception $e) {
+            return response()->json(["status" => 1, "message" => "Erreur lors du telechargement : " . $e->getMessage()], 400);
+        }
+    }
 }
